Move contents of many plugin headers to corresponding source files in TopEventProducers
List of classes:
- StEvtSolutionMaker
- TopDecaySubset
- TopInitSubset
- TtGenEventReco
- StGenEventReco
- TtSemiEvtSolutionMaker
- TtDilepEvtSolutionMaker
- TtHadEvtSolutionMaker
- TtEvtBuilder
- PseudoTopProducer
- StringCutObjectEvtFilter
