+++ TopQuarkAnalysis/TopEventProducers/src/StGenEventReco.cc
@@ -1,4 +1,20 @@
-#include "TopQuarkAnalysis/TopEventProducers/interface/StGenEventReco.h"
+#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+
+#include "AnalysisDataFormats/TopObjects/interface/StGenEvent.h"
+
+class StGenEventReco : public edm::EDProducer {
+public:
+  explicit StGenEventReco(const edm::ParameterSet&);
+  ~StGenEventReco() override;
+  void produce(edm::Event&, const edm::EventSetup&) override;
+
+private:
+  edm::EDGetTokenT<reco::GenParticleCollection> srcToken_;
+  edm::EDGetTokenT<reco::GenParticleCollection> initToken_;
+};
 
 StGenEventReco::StGenEventReco(const edm::ParameterSet& cfg)
     : srcToken_(consumes<reco::GenParticleCollection>(cfg.getParameter<edm::InputTag>("src"))),
@@ -26,3 +42,6 @@ void StGenEventReco::produce(edm::Event& evt, const edm::EventSetup& setup) {
   std::unique_ptr<StGenEvent> gen(genEvt);
   evt.put(std::move(gen));
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(StGenEventReco);
