+++ TopQuarkAnalysis/TopEventProducers/src/TopInitSubset.cc
@@ -1,4 +1,26 @@
-#include "TopQuarkAnalysis/TopEventProducers/interface/TopInitSubset.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "DataFormats/HepMCCandidate/interface/GenParticle.h"
+
+namespace TopInitID {
+  static const int status = 3;
+  static const int tID = 6;
+}  // namespace TopInitID
+
+class TopInitSubset : public edm::EDProducer {
+public:
+  explicit TopInitSubset(const edm::ParameterSet&);
+  ~TopInitSubset() override;
+
+  void produce(edm::Event&, const edm::EventSetup&) override;
+  void fillOutput(const reco::GenParticleCollection&, reco::GenParticleCollection&);
+
+private:
+  edm::EDGetTokenT<reco::GenParticleCollection> srcToken_;
+};
 
 TopInitSubset::TopInitSubset(const edm::ParameterSet& cfg)
     : srcToken_(consumes<reco::GenParticleCollection>(cfg.getParameter<edm::InputTag>("src"))) {
@@ -43,3 +65,6 @@ void TopInitSubset::fillOutput(const reco::GenParticleCollection& src, reco::Gen
     }
   }
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(TopInitSubset);
