+++ TopQuarkAnalysis/TopEventProducers/src/TtGenEventReco.cc
@@ -1,6 +1,20 @@
+#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "AnalysisDataFormats/TopObjects/interface/TtGenEvent.h"
+#include "DataFormats/HepMCCandidate/interface/GenParticleFwd.h"
 
-#include "TopQuarkAnalysis/TopEventProducers/interface/TtGenEventReco.h"
+class TtGenEventReco : public edm::EDProducer {
+public:
+  explicit TtGenEventReco(const edm::ParameterSet&);
+  ~TtGenEventReco() override;
+  void produce(edm::Event&, const edm::EventSetup&) override;
+
+private:
+  edm::EDGetTokenT<reco::GenParticleCollection> srcToken_;
+  edm::EDGetTokenT<reco::GenParticleCollection> initToken_;
+};
 
 TtGenEventReco::TtGenEventReco(const edm::ParameterSet& cfg)
     : srcToken_(consumes<reco::GenParticleCollection>(cfg.getParameter<edm::InputTag>("src"))),
@@ -28,3 +42,6 @@ void TtGenEventReco::produce(edm::Event& evt, const edm::EventSetup& setup) {
   std::unique_ptr<TtGenEvent> gen(genEvt);
   evt.put(std::move(gen));
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(TtGenEventReco);
