+++ TopQuarkAnalysis/TopEventProducers/src/TtHadEvtSolutionMaker.cc
@@ -1,7 +1,13 @@
+//
+// adapted TtSemiEvtSolutionMaker.h, v1.13 2007/07/06 02:49:42 lowette Exp $
+// for fully hadronic channel.
 
-#include "TopQuarkAnalysis/TopEventProducers/interface/TtHadEvtSolutionMaker.h"
+#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Framework/interface/Event.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
-#include "DataFormats/Math/interface/deltaR.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+
 #include "TopQuarkAnalysis/TopTools/interface/JetPartonMatching.h"
 #include "TopQuarkAnalysis/TopKinFitter/interface/TtFullHadKinFitter.h"
 #include "TopQuarkAnalysis/TopJetCombination/interface/TtHadSimpleBestJetComb.h"
@@ -10,7 +16,44 @@
 #include "TopQuarkAnalysis/TopEventSelection/interface/TtHadLRSignalSelObservables.h"
 #include "TopQuarkAnalysis/TopEventSelection/interface/TtHadLRSignalSelCalc.h"
 
+#include "DataFormats/Math/interface/deltaR.h"
+#include "AnalysisDataFormats/TopObjects/interface/TtHadEvtSolution.h"
+
 #include <memory>
+#include <string>
+#include <vector>
+
+class TtHadEvtSolutionMaker : public edm::EDProducer {
+public:
+  explicit TtHadEvtSolutionMaker(const edm::ParameterSet& iConfig);
+  ~TtHadEvtSolutionMaker() override;
+
+  void produce(edm::Event& iEvent, const edm::EventSetup& iSetup) override;
+
+private:
+  // configurables
+
+  edm::EDGetTokenT<std::vector<pat::Jet> > jetSrcToken_;
+  int jetCorrScheme_;
+  std::string lrSignalSelFile_, lrJetCombFile_;
+  bool addLRSignalSel_, addLRJetComb_, doKinFit_, matchToGenEvt_;
+  int matchingAlgo_;
+  bool useMaxDist_, useDeltaR_;
+  double maxDist_;
+  int maxNrIter_;
+  double maxDeltaS_, maxF_;
+  int jetParam_;
+  std::vector<int> lrSignalSelObs_, lrJetCombObs_;
+  std::vector<unsigned int> constraints_;
+  edm::EDGetTokenT<TtGenEvent> genEvtToken_;
+  // tools
+  TtFullHadKinFitter* myKinFitter;
+  TtHadSimpleBestJetComb* mySimpleBestJetComb;
+  TtHadLRJetCombObservables* myLRJetCombObservables;
+  TtHadLRJetCombCalc* myLRJetCombCalc;
+  TtHadLRSignalSelObservables* myLRSignalSelObservables;
+  TtHadLRSignalSelCalc* myLRSignalSelCalc;
+};
 
 /// constructor
 TtHadEvtSolutionMaker::TtHadEvtSolutionMaker(const edm::ParameterSet& iConfig) {
@@ -285,3 +328,6 @@ void TtHadEvtSolutionMaker::produce(edm::Event& iEvent, const edm::EventSetup& i
     iEvent.put(std::move(pOut));
   }
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(TtHadEvtSolutionMaker);
